requirements.txt
python-telegram-bot==21.10
httpx>=0.27
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
scipy>=1.11
apscheduler>=3.10
python-dotenv>=1.0
aiohttp>=3.9
pydantic>=2.5
rich>=13.0
